Quizz C Project Third year at ESINSA / EPUNSA (Electronic Enginiering)
The requierment where for a Graphical QUIZZ with a mandated graphic library.
Strangely, Version Control was not a thing at the time.
This is what survived a hard drive crash. I seem to remember some "test" code,
but it seems strangely, test was also not a thing at the time.